0-putchar.c

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -8,12 +8,10 @@ int main(void)
 {
 char c[] = "_putchar\n";
 int i;
-for (i = 0; i < sizeof(c) - 1; i++)
+for (i = 0; i < (int) sizeof(c) - 1; i++)
 {
 putchar(c[i]);
 }
 return (0);
 }
-
-
